requirements.txt
pytest
selenium
requests
allure-pytest
appium-python-client
python-dotenv
playwright
pytest-playwright